both jwt and keycloak way are working
